Separate add_fields and add_labels source files
