+++ libbeat/processors/actions/add_labels.go
@@ -18,33 +18,20 @@
 package actions
 
 import (
-	"encoding/json"
 	"fmt"
 
-	"github.com/elastic/beats/libbeat/beat"
 	"github.com/elastic/beats/libbeat/common"
 	"github.com/elastic/beats/libbeat/processors"
 )
 
-type addFields struct {
-	fields common.MapStr
-	shared bool
-}
-
 // LabelsKey is the default target key for the add_labels processor.
 const LabelsKey = "labels"
-const FieldsKey = "fields"
 
 func init() {
 	processors.RegisterPlugin("add_labels",
 		configChecked(createAddLabels,
 			requireFields(LabelsKey),
 			allowedFields(LabelsKey, "when")))
-
-	processors.RegisterPlugin("add_fields",
-		configChecked(createAddFields,
-			requireFields(FieldsKey),
-			allowedFields(FieldsKey, "target", "when")))
 }
 
 func createAddLabels(c *common.Config) (processors.Processor, error) {
@@ -59,40 +46,6 @@ func createAddLabels(c *common.Config) (processors.Processor, error) {
 	return makeFieldsProcessor(LabelsKey, config.Labels.Flatten(), true), nil
 }
 
-func createAddFields(c *common.Config) (processors.Processor, error) {
-	config := struct {
-		Fields common.MapStr `config:"fields" validate:"required"`
-		Target *string       `config:"target"`
-	}{}
-	err := c.Unpack(&config)
-	if err != nil {
-		return nil, fmt.Errorf("fail to unpack the add_fields configuration: %s", err)
-	}
-
-	return makeFieldsProcessor(
-		optTarget(config.Target, FieldsKey),
-		config.Fields,
-		true,
-	), nil
-}
-
-func optTarget(opt *string, def string) string {
-	if opt == nil {
-		return def
-	}
-	return *opt
-}
-
-func makeFieldsProcessor(target string, fields common.MapStr, shared bool) processors.Processor {
-	if target != "" {
-		fields = common.MapStr{
-			target: fields,
-		}
-	}
-
-	return NewAddFields(fields, shared)
-}
-
 // NewAddLabels creates a new processor adding the given object to events. Set
 // `shared` true if there is the chance of labels being changed/modified by
 // subsequent processors.
@@ -103,26 +56,4 @@ func NewAddLabels(labels common.MapStr, shared bool) processors.Processor {
 	return NewAddFields(common.MapStr{
 		LabelsKey: labels.Flatten(),
 	}, shared)
-}
-
-// NewAddFields creates a new processor adding the given fields to events.
-// Set `shared` true if there is the chance of labels being changed/modified by
-// subsequent processors.
-func NewAddFields(fields common.MapStr, shared bool) processors.Processor {
-	return &addFields{fields: fields, shared: shared}
-}
-
-func (af *addFields) Run(event *beat.Event) (*beat.Event, error) {
-	fields := af.fields
-	if af.shared {
-		fields = fields.Clone()
-	}
-
-	event.Fields.DeepUpdate(fields)
-	return event, nil
-}
-
-func (af *addFields) String() string {
-	s, _ := json.Marshal(af.fields)
-	return fmt.Sprintf("add_fields=%s", s)
 }
